feat: add fixed income type validator tests
